+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix CSS linking syntax in View.php
 - Available change database port
 - Add 'findOrThrow' method in Model.php
+- Add 'findIsset' method to Model.php
 ```
 
 **1.0.2**